dzien5/requirements.txt
altgraph==0.17.5
attrs==25.4.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.61.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librt==0.7.3
llvmlite==0.46.0
macholib==1.16.4
matplotlib==3.10.8
mypy==1.19.0
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyMatting==1.1.14
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
tifffile==2025.12.12
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
